+++ src/api/src/routes/index.ts
@@ -1,0 +1,2 @@
+export * as postsRouter from './PostRoutes';
+export * as userRouter from './UserRoutes';