+++ src/api/src/util/authentication/authenticationFunctions.ts
@@ -1,0 +1,8 @@
+import bcrypt from "bcrypt";
+
+export const generatePasswordHash = (password: string) => {
+    const password_salt = bcrypt.genSaltSync();
+	const password_hash = bcrypt.hashSync(password, password_salt);
+
+    return password_hash;
+}