+++ src/api/src/util/errors/ExtError.ts
@@ -1,0 +1,16 @@
+import HTTP_STATUS from '../../enum/HttpStatus';
+
+class ExtError extends Error{
+
+    public status : HTTP_STATUS;
+    public reason : string;
+
+    constructor(status: HTTP_STATUS, reason: string){
+        super();
+
+        this.status = status;
+        this.reason = reason;
+    }
+}
+
+export default ExtError;