+++ src/client/src/contexts/GlobalContextProvider.tsx
@@ -1,0 +1,12 @@
+import { PropsWithChildren } from 'react';
+import { BlogContextProvider } from "./BlogContext";
+
+const GlobalContextProvider = (props: PropsWithChildren) => {
+    return(
+        <BlogContextProvider> 
+            {props.children}
+        </BlogContextProvider>
+    );
+}
+
+export default GlobalContextProvider;